added new class for multicolumn partitioning
